requirements.txt
aiohttp>=3.10.0
arxiv>=2.1.3
httpx>=0.27.0
paddleocr>=2.8.1
pillow>=10.4.0
playwright>=1.51.0
pytesseract>=0.3.13
rapidfuzz>=3.9.0
